pyproject.toml
[tool.poetry]
name = "mirage-tank"
version = "0.1.0"
description = ""
authors = ["Jason Swift <anemele@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"
numpy = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
